refactor(header): split into smaller components
Use NavGroup for Swap, Earn, Analytics and About nav items. Move shared styled components into a separate file.
